Update test to include negative size. Add pr_err logs in EFAULT cases in driver. Update README to include sudo in commands.

+++ src/sma_device.cpp
@@ -5,14 +5,14 @@ SmaDevice::SmaDevice()
     fd = open("/dev/sma", O_RDWR);
 }
 
-int SmaDevice::read(uint8_t output_sequence[], int size)
+int SmaDevice::read(uint8_t output_sequence[], size_t size)
 {
     int ret;
     ret = ::read(fd, output_sequence, size);
     return ret;
 }
 
-int SmaDevice::write(uint8_t input_sequence[], int size)
+int SmaDevice::write(uint8_t input_sequence[], size_t size)
 {
     int ret;
     ret = ::write(fd, input_sequence, size);
